Add comment for using a Job rather than a bare Pod for the prober
But it's probably not worth to use a Pod. The only benefit it would
bring is the automatic deletion after completion (of the Job and the
Pod), which can be enabled with the job.spec.ttlSecondsAfterFinished
field; however, this is an alpha feature and requires the TTLAfterFinished
feature gate enabled, which cannot be assumed for most clusters.

So, the easiest solution is probably to start a bare Pod with
pod.spec.restartPolicy set to Never or OnFailure and then have the
controller delete the prober after completion (after reaching EOF in the
prober Pod output).