Add DNS tests
Testing DNS wouldn't really be necessary, as DNS resolution works
through the Service exposing the Kubernetes DNS Pods (kube-dns). Thus,
if the Service check succeeds, DNS resolution will work too, as it
consists just of accessing the DNS Pods behind the kube-dns Service.

For now, the tests are nevertheless included. If it doesn't pose too
much overhead, it's a nice-to-have.