Pin Alpine version of prober Pod to 3.11.2
This is because in newer versions, nslookup has a different behaviour:
if the DNS lookup uses the "search" suffixes in /etc/resolv.conf, and if
any of them does not succeed, the command as a whole returns a non-zero
exit code. The exit code is only 0 if all the queries (for all the
"search" suffixes) succeed, which is usually not the case.

nslookup in Alpine 3.11.2 returns 0 if any of the queries succeeds, that
is, if the name can actually be resolved, and a non-zero exit code only
if no query at all sucdeeds. This is the desired behaviour.

See https://github.com/gliderlabs/docker-alpine/issues/539