Turn prober Pods into Deployments
In Deployments, the list of environment variable of any container can be
patched (in contrast to Pods). This allows the prober Pods to be created
with all the other resources (in conncheck.yaml) and the controller Pod
then adds the missing environment variables with the topology info to
them. After patching a Deployment, the Pods get restarted by the
Deployment controller, thus the prober Pods will see the new environment
variables immediately when they start up (they only start running the
tests when these environment variables exist, otherwise they just idle).

The goal of all this is to keep the prober Pods running after they
complete the tests, which is useful for investigations when tests fail.
By being able to include the prober Pods in the conncheck.yaml manifest,
the manifest is self-contained.

Furthermore, the prober Pods being managed as Deployments has certain
advantages for future features. For example, it could be possible to
delete a prober Pod, in which case it's restarted by the Deployment and
automatically starts running the tests again (for repeating the tests).