Prober Deployments in root manifest patched by controller Pod
The Deployments with the prober Pods are created initially. Then, after
the controller Pod gathered the topology information, it patches the
Deployments, adding the topology information as environment variables to
the Pod template. This causes the Deployment to restart the prober Pods,
in which case they start running the tests.

Advantages:

- All created resources are in the root manifest (conncheck.yaml)

Disadvantages:

- Bad UX: right after launching the application, two Pods are
Terminating. It takes quite a while until the terminating Pods are
completely terminated and deleted.
- The prober Deployments as defined in the root manifest are not
complete because they will be patched later by the controller Pod.