Controller creates prober Deployments; prober Pod output results
Revert to the controller Pod creating the prober Deployments, rather
then them being defined in the conncheck.yaml manifest and then patched
by the controller Pod (see disadvantages of this approach in commit
message of previous commit).

Additonally, the user-facing results are now output directly by the
prober Pods, rather than being collected and output by the controller
Pod as before. This removes a lot of implementation complexity while
maintaining a comparable UX.

To view the test results, the user can do:

  kubectl logs deployment/conncheck-prober
  kubectl logs deployment/conncheck-prober-hostnet

Instead of:

  kubectl logs conncheck-controller

as before.