Extract API server IP address from iptables rules
Drill down into the iptables rules for the "default/kubernetes" Service
which exists in every cluster and exposes the API server for internal
cluster users. These rules contain the destination IP address and port
that kube-proxy configured for DNAT'ing requests to the ClusterIP of the
"default/kubernetes" Service.

This makes the application self-configurable (no need to manually
specify the API server URL), but poses several requirements:

- Requires the container of the controller Pod to run in privileded
mode (i.e. can't be used in a cluster that doesn't allow containers in
privileged mode).
- Probably works only if kube-proxy uses the 'iptables' proxy mode. In
the IPVS proxy mode, kube-proxy probably doesn't set iptables rules.
However, this has not yet been tested.

So far, the functionality has been successfully tested on a cluster
created with kubeadm and on GKE.