Check if prober ReplicaSets are already running; if yes, patch them
The intention was to allow restarting the controller Pod to gather the
topology information again (useful in case it has been changed) and then
patching the existing prober ReplicaSets with the new topology
information, which would cause the prober Pods to be restarted.

However, this doesn't work as expected:

- The Pods managed by a ReplicaSet are never restarted when patching the
ReplicaSet.
- With a Deployment, the Pods are restarted when patching the
Deployment, but only when the patch actually changes the Deployment
definition (in particular, the Pod template). If the patch doesn't
change anything, the Pods are not restarted.

To have a deterministic behaviour, the prober Pods should always restart
when restarting the controller Pod. That means, the Pods would need to
be restarted manually anyway, regardless of whether a Deployment or a
ReplicaSet is used.

For this reason, it's probably better to use a different approach:
instead of using `kubectl create` to create the ReplicaSets and `kubectl
patch` to patch them, use `kubectl apply` for both cases, and in the
case that the ReplicaSets already exist, restart the Pods.

Note: `kubectl apply` does not result in the Pods being restarted, even
if the Pod template of the corresponding Deployment or ReplicaSet changes.